+++ DependencyInjection/Configuration.php
@@ -1,13 +1,12 @@
 <?php
-
 namespace Oka\ApiBundle\DependencyInjection;
 
+use Oka\ApiBundle\CorsOptions;
+use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
+use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\Config\Definition\Builder\ScalarNodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Oka\ApiBundle\CorsOptions;
-use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 
 /**
  * This is the class that validates and merges configuration from your app/config files.
@@ -57,6 +56,12 @@ class Configuration implements ConfigurationInterface
 // 							->append($this->getJWTFirewallNodeDefintion('jwt'))
 						->end()
 						->info('This value configure API firewalls.')
+					->end()
+					->arrayNode('response')
+						->addDefaultsIfNotSet()
+						->children()
+							->scalarNode('error_builder_class')->defaultNull()->end()
+						->end()
 					->end()
 				->end();