+++ Http/HostRequestMatcher.php
@@ -1,8 +1,8 @@
 <?php
 namespace Oka\ApiBundle\Http;
 
-use Symfony\Component\HttpFoundation\RequestMatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestMatcherInterface;
 
 /**
  *